+++ python/jaeger/target/design.py
@@ -24,9 +24,26 @@ __all__ = ["Design"]
 
 
 class Design:
-    """Loads and represents a targetdb design."""
+    """Loads and represents a targetdb design.
 
-    def __init__(self, design_id: int, load_configuration=True):
+    Parameters
+    ----------
+    design_id
+        The ID of the design to load.
+    load_configuration
+        Create a `.Configuration` attached to this design.
+    epoch
+        The JD epoch for which to calculate the configuration coordinates. If
+        `None`, uses the current time.
+
+    """
+
+    def __init__(
+        self,
+        design_id: int,
+        load_configuration: bool = True,
+        epoch: float | None = None,
+    ):
 
         if calibration.wokCoords is None:
             raise RuntimeError("Cannot retrieve wok calibration. Is $WOKCALIB_DIR set?")
@@ -50,7 +67,7 @@ class Design:
 
         self.configuration: Configuration
         if load_configuration:
-            self.configuration = Configuration(self)
+            self.configuration = Configuration(self, epoch=epoch)
 
     def get_target_data(self) -> dict[str, dict]:
         """Retrieves target data as a dictionary."""
@@ -97,12 +114,12 @@ class Design:
         return {data["holeid"]: data for data in target_data}
 
     @classmethod
-    async def create_async(cls, design_id: int):
+    async def create_async(cls, design_id: int, epoch: float | None = None):
         """Returns a design while creating the configuration in an executor."""
 
         self = cls(design_id, load_configuration=False)
 
-        configuration = await run_in_executor(Configuration, self)
+        configuration = await run_in_executor(Configuration, self, epoch=epoch)
         self.configuration = configuration
 
         return self